Upgrade to docusaurus 3.2.1
Should fix many of the security issues with JS dependencies pointed
out by Github Dependabot.

Also, don't hide inside package.json how we start the development
environment. The default setup of docusaurus doesn't do that. If we
want to start the dev environment in a particular way, make it
explicit in the docker-compose.yml file, so everybody can know we are
doing special things.